move pager routines to `show_messsage.c`
Also:
* Remove some unnecessary includes.
