+++ handlers.c
@@ -30,13 +30,10 @@
 
 #include <stdio.h>
 #include <sys/types.h>
-#include <sys/wait.h>
 #include <sys/ioctl.h>
-#include <fcntl.h>
 #include <unistd.h>
 #include <stdlib.h>
 #include <getopt.h>
-#include <signal.h>
 #include <assert.h>
 #include <setjmp.h>
 #include <alloca.h>
@@ -90,109 +87,6 @@
 #else
 #define POINTER_FMT "%12lx"
 #endif
-
-/* pager support */
-static FILE *get_pager(FILE *fallback_output)
-{
-    const char *pager;
-    pid_t pgpid;
-    int pgret;
-    int pgpipe[2];
-    bool pgcmdfail = false;
-    FILE *retfd = NULL;
-    char *const emptyvec[1] = { NULL };
-
-    assert(fallback_output != NULL);
-
-    if (sm_globals.options.backend)
-        return fallback_output;
-
-    if ((pager = util_getenv("PAGER")) == NULL || *pager == '\0') {
-        show_warn("get_pager(): couldn't get $PAGER, falling back to `more`\n");
-        pager = "more";
-    }
-
-    if (pipe2(pgpipe, O_NONBLOCK) == -1) {
-        show_error("get_pager(): pipe2() error `%s`. falling back to normal output\n", strerror(errno));
-        return fallback_output;
-    }
-    /*
-     * we write here to ensure we will always
-     * have something to read() into pgcmdfail.
-     */
-    write(pgpipe[1], "", 1);
-
-    /* XXX: is $PATH modified prior? */
-retry:
-    switch ((pgpid = fork())) {
-    case -1:
-        show_warn("get_pager(): fork() failed. falling back to normal output\n");
-        return fallback_output;
-    case 0:
-        execvp(pager, emptyvec);
-        /*
-         * if we got here, it means execvp() failed.
-         * errno contains the error, so pass it back
-         * up to parent. we use a pipe here to let
-         * the parent know that we are indeed returning
-         * the return value of the failed execvp().
-         */
-        char nullbuf;
-        /* read() to empty pipe */
-        read(pgpipe[0], &nullbuf, 1);
-        write(pgpipe[1], "1", 2);
-        exit(errno);
-        /* NOTREACHED */
-    default:
-        if (waitpid(pgpid, &pgret, 0) == -1) {
-            show_debug("pager: waitpid() error `%s`\n", strerror(errno));
-            show_warn("pager: waitpid() error. falling back to normal output\n");
-            return fallback_output;
-        } else {
-            pgret = WEXITSTATUS(pgret);
-            if (read(pgpipe[0], &pgcmdfail, 1) == -1) {
-                show_error("pager: pipe read() error `%s`. falling back to normal output\n", strerror(errno));
-                return fallback_output;
-            }
-            if (pgcmdfail) {
-                show_debug("pager: execvp(pg=%s) ret -> %d (%s)\n", pager, pgret, strerror(pgret));
-                if (!strcmp(pager, "more")) {
-                    show_warn("pager: sh failed to execute more. falling back to normal output\n");
-                    return fallback_output;
-                } else {
-                    show_warn("pager: sh failed to execute `%s`. trying `more`...\n", pager);
-                    pager = "more";
-                    goto retry;
-                }
-            }
-        }
-    }
-
-    if ((retfd = popen(pager, "w")) == NULL) {
-        show_warn("pager: couldn't popen() pager, falling back to normal output\n");
-        return fallback_output;
-    }
-
-    /*
-     * we need to ignore SIGPIPE in case the pager exits wihout draining the
-     * pipe (less seems to do this if you exit a large listing without
-     *       scrolling down)
-     */
-    signal(SIGPIPE, SIG_IGN);
-
-    assert(retfd != NULL);
-
-    return retfd;
-}
-
-static inline void close_pager(FILE *pager)
-{
-    if (pager != stdout && pager != stderr) {
-        if (pclose(pager) == -1 && errno != EPIPE)
-            show_warn("pclose() error: %s\n", strerror(errno));
-        signal(SIGPIPE, SIG_DFL);
-    }
-}
 
 bool handler__set(globals_t * vars, char **argv, unsigned argc)
 {
